리팩토링/refactoring/src/main/java/me/whiteship/refactoring/_18_middle_man/_40_replace_subclass_with_delegate/_02_after/Booking.java
package me.whiteship.refactoring._18_middle_man._40_replace_subclass_with_delegate._02_after;

import java.time.DayOfWeek;
import java.time.LocalDateTime;

public class Booking {

    protected Show show;

    protected LocalDateTime time;

    protected PremiumDelegate premiumDelegate;

    public Booking(Show show, LocalDateTime time) {
        this.show = show;
        this.time = time;
    }

    /**
     * 아래와 같이 static 한 팩토리 메서드를 만들때는 얼마든지 이름을 자유롭게 표현할수 있음!
     *
     * @param show
     * @return
     */
    public static Booking createBook(Show show, LocalDateTime time) {
        return new Booking(show, time);
    }

    public static Booking createPremiumBooking(Show show, LocalDateTime time, PremiumExtra extra) {

        /**
         * 여기서 좋은 점이 PremiumBooking  객체를 생성할 필요가 없음
         */
        Booking booking = createBook(show, time);
        //PremiumBooking booking = new PremiumBooking(show, time, extra);
        booking.premiumDelegate = new PremiumDelegate(booking, extra);
        return booking;
    }

    public boolean hasTalkback() {
        return (this.premiumDelegate != null) ? this.premiumDelegate.hasTalkback() :
                this.show.hasOwnProperty("talkback") && !this.isPeakDay();
    }

    protected boolean isPeakDay() {
        DayOfWeek dayOfWeek = this.time.getDayOfWeek();
        return dayOfWeek == DayOfWeek.SATURDAY || dayOfWeek == DayOfWeek.SUNDAY;
    }

    public double basePrice() {
        double result = this.show.getPrice();
        if (this.isPeakDay()) result += Math.round(result * 0.15);
        return (this.premiumDelegate != null) ? this.premiumDelegate.extendBasePrice(result) : result;
    }

    public boolean hasDinner() {
        return this.premiumDelegate != null && this.premiumDelegate.hasDinner();
    }
}
